+++ lib/sql_beautifier/clauses/base.rb
@@ -10,6 +10,16 @@ module SqlBeautifier
       def initialize(value)
         @value = value
       end
+
+      private
+
+      def keyword_prefix
+        Util.keyword_padding(self.class::KEYWORD)
+      end
+
+      def continuation_indent
+        Util.continuation_padding
+      end
     end
   end
 end