+++ lib/sql_beautifier/clauses/group_by.rb
@@ -3,10 +3,10 @@
 module SqlBeautifier
   module Clauses
     class GroupBy < Base
-      KEYWORD_PREFIX = "group by "
+      KEYWORD = "group by"
 
       def call
-        "#{KEYWORD_PREFIX}#{@value.strip}"
+        "#{keyword_prefix}#{@value.strip}"
       end
     end
   end